Restructure lesson viewer files to be shared
