+++ frontend/src/components/module-viewer/BlockPreview.tsx
