+++ frontend/src/components/module-viewer/EditContentOptionsMenu.tsx
