+++ frontend/src/components/module-viewer/LessonItem.tsx
