+++ frontend/src/components/module-viewer/SideBar.tsx
