+++ frontend/src/components/module-viewer/SideBarContentKiosk.tsx
