+++ frontend/src/components/module-viewer/SideBarModuleOverview.tsx
