+++ frontend/src/components/module-viewer/content/modals/EditImageModal.tsx
